XMLHttpRequest/tests/submissions/Opera/send-authentication.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - "Basic" authenticated requests</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function request(user1, pass1, user2, pass2) {
        test(function() {
          var client = new XMLHttpRequest(),
              urlstart = "",
              userwin = user2 || user1,
              passwin = pass2 || pass1
          if(user1) {
            urlstart = "http://" + user1
            if(pass1)
              urlstart += ":" + pass1
            urlstart += "@localhost/xhr/"
          }
          client.open("GET", urlstart + "resources/authentication.php", false, user2, pass2)
          client.setRequestHeader("x-user", userwin)
          client.setRequestHeader("x-pass", passwin)
          client.send(null)
          assert_equals(client.responseText, userwin + "\n" + passwin)
        }, document.title)
      }
      var ident = Math.random()
      request(null, null, ident+1, ident+2)
      request(null, null, ident+3, ident+4)
      request("1", "2", ident+5, ident+6)
      request(ident+7, ident+8, null, null)
      request(ident+9, null, null, ident+10)
      request("1", ident+11, ident+12, null)
    </script>
  </body>
</html>
